+++ packages/adapter-static/CHANGELOG.md
@@ -1,5 +1,19 @@
 # houdini-adapter-static
 
+## 2.0.0-go.16
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   houdini@2.0.0-go.16
+
+## 2.0.0-go.15
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   houdini@2.0.0-go.15
+
 ## 2.0.0-go.14
 
 ### Patch Changes